# Backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UnnamedMessenger API",
    "description": "API specification for secure messenger with E2E encryption",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/chats": {
      "get": {
        "summary": "Get list of user's chats",
        "tags": ["Chats"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": ["pm", "gm"]
                      },
                      "name": {
                        "type": "string"
                      },
                      "members": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "lastMessage": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "text": {
                            "type": "string"
                          },
                          "timestamp": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new chat",
        "tags": ["Chats"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type", "encryptedKey"],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["pm", "gm"]
                  },
                  "name": {
                    "type": "string"
                  },
                  "members": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "encryptedKey": {
                    "type": "string",
                    "description": "Chat key encrypted with creator's public key"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Chat created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/chats/{chatId}": {
      "get": {
        "summary": "Get chat by ID",
        "tags": ["Chats"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Chat not found"
          }
        }
      },
      "delete": {
        "summary": "Delete chat",
        "tags": ["Chats"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not allowed to delete this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        }
      }
    },
    "/chats/{chatId}/members/add": {
      "post": {
        "summary": "Add member to chat",
        "tags": ["Chats"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId", "encryptedKey"],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "encryptedKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Member added successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not allowed to add members"
          },
          "404": {
            "description": "Chat not found"
          }
        }
      }
    },
    "/chats/{chatId}/members/remove": {
      "post": {
        "summary": "Remove member from chat",
        "tags": ["Chats"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["userId"],
                "properties": {
                  "userId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member removed successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not allowed to remove members"
          },
          "404": {
            "description": "Chat or user not found"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  }
}